requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.7.0
aptos-sdk==0.10.0
async-timeout==5.0.1
attrs==24.3.0
behave==1.2.6
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
ecdsa==0.18.0
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
multidict==6.1.0
oauthlib==3.2.2
parse==1.20.2
parse_type==0.6.4
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
python-dotenv==1.0.1
python-graphql-client==0.4.3
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
tomli==2.2.1
types-six==1.17.0.20241205
typing_extensions==4.12.2
urllib3==2.3.0
websockets==14.1
yarl==1.18.3
